Move JiraTicketRow out of BBComponents
